Completed both parts of day 6.  Had to convert Answer to a long.

+++ AoC_2021/ChallengeBase.cs
@@ -16,5 +16,5 @@ public class ChallengeBase : IChallenge
     {
     }
 
-    public virtual int Answer => throw new NotImplementedException();
+    public virtual long Answer => throw new NotImplementedException();
 }
